pyproject.toml
[project]
name = "appian-mcp-server"
version = "0.1.0"
description = "MCP Server for Appian CSR Request API"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]
